requirements.txt
torch>=2.5.0
torchvision
numpy<2
opencv-python
timm
scipy
scikit-image
kornia
einops
mediapipe==0.10.14
pillow

fastapi
uvicorn
cloudinary
python-dotenv

tqdm
prettytable
tabulate
ipykernel

huggingface-hub>0.25
accelerate
